# openapi/product.json
{
  "components": {
    "schemas": {
...
      "CommandRunView": {
        "additionalProperties": false,
        "properties": {
          "cancel_action": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ProductAction"
              },
              {
                "type": "null"
              }
            ]
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "elapsed_seconds": {
            "anyOf": [
              {
                "minimum": 0.0,
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Elapsed Seconds"
          },
          "ended_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Ended At"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "member": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/TaskMemberReference"
              },
              {
                "type": "null"
              }
            ]
          },
          "outcome_summary": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Outcome Summary"
          },
          "output_complete": {
            "title": "Output Complete",
            "type": "boolean"
          },
          "output_href": {
            "title": "Output Href",
            "type": "string"
          },
          "purpose": {
            "title": "Purpose",
            "type": "string"
          },
          "started_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Started At"
          },
          "state": {
            "$ref": "#/components/schemas/CommandRunProductState"
          }
        },
        "required": [
          "id",
          "purpose",
          "state",
          "created_at",
          "output_href",
          "output_complete"
        ],
        "title": "CommandRunView",
        "type": "object"
      },
      "FileReference": {
        "additionalProperties": false,
        "description": "An ordered navigation value to one ordinary workspace file.",
        "properties": {
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "path": {
            "title": "Path",
            "type": "string"
          }
        },
        "required": [
          "path"
        ],
        "title": "FileReference",
        "type": "object"
      },
...
      "JsonValue": {},
...
      "OperationFailure": {
        "additionalProperties": false,
        "properties": {
          "code": {
            "$ref": "#/components/schemas/ProductFailureCode"
          },
          "field_path": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Field Path"
          },
          "ok": {
            "const": false,
            "default": false,
            "title": "Ok",
            "type": "boolean"
          },
          "retryable": {
            "title": "Retryable",
            "type": "boolean"
          },
          "suggested_next_step": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Suggested Next Step"
          },
          "summary": {
            "title": "Summary",
            "type": "string"
          }
        },
        "required": [
          "code",
          "summary",
          "retryable"
        ],
        "title": "OperationFailure",
        "type": "object"
      },
      "ProductAction": {
        "additionalProperties": false,
        "properties": {
          "confirmation": {
            "$ref": "#/components/schemas/ProductActionConfirmation"
          },
          "href": {
            "title": "Href",
            "type": "string"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "input_schema": {
            "anyOf": [
              {
                "additionalProperties": {
                  "$ref": "#/components/schemas/JsonValue"
                },
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Input Schema"
          },
          "kind": {
            "title": "Kind",
            "type": "string"
          },
          "label": {
            "title": "Label",
            "type": "string"
          }
        },
        "required": [
          "id",
          "kind",
          "label",
          "href",
          "confirmation"
        ],
        "title": "ProductAction",
        "type": "object"
      },
...
      "ProductFailureCode": {
        "description": "Closed, product-safe failure vocabulary shared by HTTP and Operator.",
        "enum": [
          "invalid_request",
          "not_found",
          "conflict",
          "cursor_reset_required",
          "access_denied",
          "unavailable",
          "internal_error"
        ],
        "title": "ProductFailureCode",
        "type": "string"
      },
...
      "TaskActivity": {
        "additionalProperties": false,
        "properties": {
          "action": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/TaskActivityLink"
              },
              {
                "type": "null"
              }
            ]
          },
          "files": {
            "default": [],
            "items": {
              "$ref": "#/components/schemas/FileReference"
            },
            "title": "Files",
            "type": "array"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "kind": {
            "$ref": "#/components/schemas/TaskActivityKind"
          },
          "member": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/TaskMemberReference"
              },
              {
                "type": "null"
              }
            ]
          },
          "occurred_at": {
            "format": "date-time",
            "title": "Occurred At",
            "type": "string"
          },
          "outcome": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/TaskActivityOutcome"
              },
              {
                "type": "null"
              }
            ]
          },
          "summary": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Summary"
          },
          "title": {
            "title": "Title",
            "type": "string"
          }
        },
        "required": [
          "id",
          "kind",
          "occurred_at",
          "title"
        ],
        "title": "TaskActivity",
        "type": "object"
      },
...
      "TaskAttention": {
        "additionalProperties": false,
        "properties": {
          "action": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ProductAction"
              },
              {
                "type": "null"
              }
            ]
          },
          "files": {
            "default": [],
            "items": {
              "$ref": "#/components/schemas/FileReference"
            },
            "title": "Files",
            "type": "array"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "kind": {
            "$ref": "#/components/schemas/TaskAttentionKind"
          },
          "link": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/TaskActivityLink"
              },
              {
                "type": "null"
              }
            ]
          },
          "member": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/TaskMemberReference"
              },
              {
                "type": "null"
              }
            ]
          },
          "summary": {
            "title": "Summary",
            "type": "string"
          },
          "title": {
            "title": "Title",
            "type": "string"
          }
        },
        "required": [
          "id",
          "kind",
          "title",
          "summary"
        ],
        "title": "TaskAttention",
        "type": "object"
      },
...
      "TaskControlKind": {
        "enum": [
          "pause",
          "resume",
          "cancel"
        ],
        "type": "string"
      },
      "TaskControlReceipt": {
        "additionalProperties": false,
        "properties": {
          "action": {
            "$ref": "#/components/schemas/TaskControlKind"
          },
          "receipt_id": {
            "title": "Receipt Id",
            "type": "string"
          },
          "status_message": {
            "title": "Status Message",
            "type": "string"
          },
          "task": {
            "$ref": "#/components/schemas/TaskView"
          }
        },
        "required": [
          "receipt_id",
          "action",
          "status_message",
          "task"
        ],
        "title": "TaskControlReceipt",
        "type": "object"
      },
      "TaskControlRequest": {
        "additionalProperties": false,
        "properties": {
          "confirmed": {
            "default": false,
            "title": "Confirmed",
            "type": "boolean"
          }
        },
        "title": "TaskControlRequest",
        "type": "object"
      },
...
      "TaskPlanView": {
        "additionalProperties": false,
        "properties": {
          "explanation": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Explanation"
          },
          "steps": {
            "items": {
              "$ref": "#/components/schemas/TaskPlanStep"
            },
            "title": "Steps",
            "type": "array"
          },
          "updated_at": {
            "format": "date-time",
            "title": "Updated At",
            "type": "string"
          }
        },
        "required": [
          "steps",
          "updated_at"
        ],
        "title": "TaskPlanView",
        "type": "object"
      },
      "TaskProductStatus": {
        "enum": [
          "starting",
          "working",
          "waiting_for_you",
          "paused",
          "completed",
          "blocked",
          "cancelled"
        ],
        "type": "string"
      },
...
      "TaskSearchResponse": {
        "additionalProperties": false,
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/TaskSummary"
            },
            "title": "Items",
            "type": "array"
          },
          "next_cursor": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Next Cursor"
          }
        },
        "required": [
          "items"
        ],
        "title": "TaskSearchResponse",
        "type": "object"
      },
      "TaskStartReceipt": {
        "additionalProperties": false,
        "properties": {
          "manifest": {
            "title": "Manifest",
            "type": "string"
          },
          "receipt_id": {
            "title": "Receipt Id",
            "type": "string"
          },
          "status": {
            "const": "accepted",
            "default": "accepted",
            "title": "Status",
            "type": "string"
          },
          "status_message": {
            "default": "The run was accepted. Work starts asynchronously and may still need attention.",
            "title": "Status Message",
            "type": "string"
          },
          "task_id": {
            "title": "Task Id",
            "type": "string"
          },
          "workflow_id": {
            "title": "Workflow Id",
            "type": "string"
          },
          "workflow_revision": {
            "minimum": 1.0,
            "title": "Workflow Revision",
            "type": "integer"
          },
          "workspace": {
            "title": "Workspace",
            "type": "string"
          }
        },
        "required": [
          "receipt_id",
          "task_id",
          "workflow_id",
          "workflow_revision",
          "workspace",
          "manifest"
        ],
        "title": "TaskStartReceipt",
        "type": "object"
      },
      "TaskStartRequest": {
        "additionalProperties": false,
        "description": "The one transient request accepted by every Task-start transport.",
        "properties": {
          "files": {
            "default": [],
            "items": {
              "$ref": "#/components/schemas/FileReference"
            },
            "title": "Files",
            "type": "array"
          },
          "prompt": {
            "title": "Prompt",
            "type": "string"
          },
          "workflow": {
            "maxLength": 128,
            "minLength": 1,
            "pattern": "^[a-z][a-z0-9]*(?:[-_][a-z0-9]+)*$",
            "title": "Workflow",
            "type": "string"
          },
          "workspace": {
            "anyOf": [
              {
                "format": "path",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Workspace"
          }
        },
        "required": [
          "workflow",
          "prompt"
        ],
        "title": "TaskStartRequest",
        "type": "object"
      },
      "TaskSummary": {
        "additionalProperties": false,
        "properties": {
          "attention_count": {
            "minimum": 0.0,
            "title": "Attention Count",
            "type": "integer"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "prompt_excerpt": {
            "title": "Prompt Excerpt",
            "type": "string"
          },
          "result_status": {
            "anyOf": [
              {
                "enum": [
                  "completed",
                  "blocked"
                ],
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Result Status"
          },
          "started_at": {
            "format": "date-time",
            "title": "Started At",
            "type": "string"
          },
          "status": {
            "$ref": "#/components/schemas/TaskProductStatus"
          },
          "status_message": {
            "title": "Status Message",
            "type": "string"
          },
          "updated_at": {
            "format": "date-time",
            "title": "Updated At",
            "type": "string"
          },
          "workflow": {
            "$ref": "#/components/schemas/TaskWorkflowView"
          }
        },
        "required": [
          "id",
          "prompt_excerpt",
          "workflow",
          "status",
          "status_message",
          "started_at",
          "updated_at",
          "attention_count"
        ],
        "title": "TaskSummary",
        "type": "object"
      },
      "TaskView": {
        "additionalProperties": false,
        "properties": {
          "actions": {
            "default": [],
            "items": {
              "$ref": "#/components/schemas/ProductAction"
            },
            "title": "Actions",
            "type": "array"
          },
          "activities": {
            "default": [],
            "items": {
              "$ref": "#/components/schemas/TaskActivity"
            },
            "title": "Activities",
            "type": "array"
          },
          "activities_href": {
            "title": "Activities Href",
            "type": "string"
          },
          "activities_truncated": {
            "default": false,
            "title": "Activities Truncated",
            "type": "boolean"
          },
          "attention": {
            "default": [],
            "items": {
              "$ref": "#/components/schemas/TaskAttention"
            },
            "title": "Attention",
            "type": "array"
          },
          "command_run_count": {
            "default": 0,
            "minimum": 0.0,
            "title": "Command Run Count",
            "type": "integer"
          },
          "command_runs": {
            "default": [],
            "items": {
              "$ref": "#/components/schemas/CommandRunView"
            },
            "title": "Command Runs",
            "type": "array"
          },
          "command_runs_truncated": {
            "default": false,
            "title": "Command Runs Truncated",
            "type": "boolean"
          },
          "human_request_count": {
            "default": 0,
            "minimum": 0.0,
            "title": "Human Request Count",
            "type": "integer"
          },
          "human_requests": {
            "default": [],
            "items": {
              "$ref": "#/components/schemas/HumanRequestView"
            },
            "title": "Human Requests",
            "type": "array"
          },
          "human_requests_truncated": {
            "default": false,
            "title": "Human Requests Truncated",
            "type": "boolean"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "plan": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/TaskPlanView"
              },
              {
                "type": "null"
              }
            ]
          },
          "prompt_excerpt": {
            "title": "Prompt Excerpt",
            "type": "string"
          },
          "result": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/TaskResultView"
              },
              {
                "type": "null"
              }
            ]
          },
          "started_at": {
            "format": "date-time",
            "title": "Started At",
            "type": "string"
          },
          "status": {
            "$ref": "#/components/schemas/TaskProductStatus"
          },
          "status_message": {
            "title": "Status Message",
            "type": "string"
          },
          "team": {
            "$ref": "#/components/schemas/TaskMemberView"
          },
          "updated_at": {
            "format": "date-time",
            "title": "Updated At",
            "type": "string"
          },
          "workflow": {
            "$ref": "#/components/schemas/TaskWorkflowView"
          }
        },
        "required": [
          "id",
          "prompt_excerpt",
          "workflow",
          "status",
          "status_message",
          "started_at",
          "updated_at",
          "team",
          "activities_href"
        ],
        "title": "TaskView",
        "type": "object"
      },
      "TaskWorkflowView": {
        "additionalProperties": false,
        "properties": {
          "description": {
            "title": "Description",
            "type": "string"
          },
          "id": {
            "title": "Id",
            "type": "string"
          }
        },
        "required": [
          "id",
          "description"
        ],
        "title": "TaskWorkflowView",
        "type": "object"
      },
...
    }
  },
  "info": {
    "title": "Banksia Product API",
    "version": "0.0.0"
  },
  "openapi": "3.1.0",
  "paths": {
    "/tasks": {
      "get": {
        "operationId": "get_tasks_tasks_get",
        "parameters": [
          {
            "in": "query",
            "name": "q",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Q"
            }
          },
          {
            "in": "query",
            "name": "status",
            "required": false,
            "schema": {
              "default": "any",
              "title": "Status",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "cursor",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "minLength": 1,
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Cursor"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 50,
              "maximum": 100,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TaskSearchResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OperationFailure"
                }
              }
            },
            "description": "Bad Request"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OperationFailure"
                }
              }
            },
            "description": "Forbidden"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OperationFailure"
                }
              }
            },
            "description": "Not Found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OperationFailure"
                }
              }
            },
            "description": "Conflict"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OperationFailure"
                }
              }
            },
            "description": "Unprocessable Content"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OperationFailure"
                }
              }
            },
            "description": "Internal Server Error"
          }
        },
        "summary": "Get Tasks",
        "tags": [
          "tasks"
        ]
      },
      "post": {
        "operationId": "post_task_tasks_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TaskStartRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "202": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TaskStartReceipt"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OperationFailure"
                }
              }
            },
            "description": "Bad Request"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OperationFailure"
                }
              }
            },
            "description": "Forbidden"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OperationFailure"
                }
              }
            },
            "description": "Not Found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OperationFailure"
                }
              }
            },
            "description": "Conflict"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OperationFailure"
                }
              }
            },
            "description": "Unprocessable Content"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OperationFailure"
                }
              }
            },
            "description": "Internal Server Error"
          }
        },
        "summary": "Post Task",
        "tags": [
          "tasks"
        ]
      }
    },
    "/tasks/{task_id}": {
      "get": {
        "operationId": "get_task_tasks__task_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "task_id",
            "required": true,
            "schema": {
              "title": "Task Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TaskView"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OperationFailure"
                }
              }
            },
            "description": "Bad Request"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OperationFailure"
                }
              }
            },
            "description": "Forbidden"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OperationFailure"
                }
              }
            },
            "description": "Not Found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OperationFailure"
                }
              }
            },
            "description": "Conflict"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OperationFailure"
                }
              }
            },
            "description": "Unprocessable Content"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OperationFailure"
                }
              }
            },
            "description": "Internal Server Error"
          }
        },
        "summary": "Get Task",
        "tags": [
          "tasks"
        ]
      }
    },
...
    "/tasks/{task_id}/controls/{action_id}": {
      "post": {
        "operationId": "post_task_control_tasks__task_id__controls__action_id__post",
        "parameters": [
          {
            "in": "path",
            "name": "task_id",
            "required": true,
            "schema": {
              "title": "Task Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "action_id",
            "required": true,
            "schema": {
              "title": "Action Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TaskControlRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TaskControlReceipt"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OperationFailure"
                }
              }
            },
            "description": "Bad Request"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OperationFailure"
                }
              }
            },
            "description": "Forbidden"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OperationFailure"
                }
              }
            },
            "description": "Not Found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OperationFailure"
                }
              }
            },
            "description": "Conflict"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OperationFailure"
                }
              }
            },
            "description": "Unprocessable Content"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OperationFailure"
                }
              }
            },
            "description": "Internal Server Error"
          }
        },
        "summary": "Post Task Control",
        "tags": [
          "tasks"
        ]
      }
    },
...
  }
}